requirements.txt
Pillow==9.5.0
rasterio==1.3.6
